pyproject.toml
[tool.poetry]
name = "authentication-service"
version = "0.1.0"
description = "Authentication Microservice for distributed systems"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11.3"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
alembic = "^1.13.1"
SQLAlchemy = "^2.0.0"
httpx = "^0.25.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "."
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = ["passlib.*", "jose.*", "pydantic_settings.*", "pydantic.*"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''